Updating easyvk, changing atomic rmw to refer to all buffer sizes as u64's

+++ benchmarks/atomic-rmw/atomic_rmw_test.cpp
@@ -54,8 +54,8 @@ extern "C" void rmw_microbenchmark(easyvk::Device device, uint32_t workgroups, u
 
             benchmark_data << "(" << contention << ", " << padding << ", ";
 
-            uint32_t global_work_size = workgroup_size * workgroups;
-            uint32_t size = ((global_work_size) * padding) / contention;
+            uint64_t global_work_size = workgroup_size * workgroups;
+            uint64_t size = ((global_work_size) * padding) / contention;
 
             Buffer result_buf = Buffer(device, (thread_dist == "random_access" ? contention : size) * sizeof(uint32_t), true);
